// src/shared/enums/HistoryEventType.ts
export enum HistoryEventType {
  ORDER_CREATED = 'ORDER_CREATED',
  ORDER_PAYMENT_SEND_BY_LINK = 'ORDER_PAYMENT_SEND_BY_LINK',
  ORDER_CHECKOUT_SESSION_CREATED = 'ORDER_CHECKOUT_SESSION_CREATED',
  ORDER_PAID_ONLINE_BY_CREDIT_CARD = 'ORDER_PAID_ONLINE_BY_CREDIT_CARD',
  DEPOSIT_PAID_ONLINE_BY_CREDIT_CARD = 'DEPOSIT_PAID_ONLINE_BY_CREDIT_CARD',
  ORDER_PAID_BY_GIFT_CARD = 'ORDER_PAID_BY_GIFT_CARD',
  ORDER_ZERO_PAID = 'ORDER_ZERO_PAID',
  ORDER_COMPLETED = 'ORDER_COMPLETED',
  ORDER_CANCELED = 'ORDER_CANCELED',
  ORDER_CHECKOUT_SESSION_EXPIRED = 'ORDER_CHECKOUT_SESSION_EXPIRED',
  ORDER_REOPENED = 'ORDER_REOPENED',
}